requirements.txt
fastapi
uvicorn
requests
flask
psycopg2-binary
